fix: Cleanup types and type assumptions. May need to revisit errors. Need to resolve mux-audio
