+++ packages/mux-player/src/types.d.ts
@@ -1,4 +1,5 @@
-import type MuxVideoElement, { MediaError } from '@mux/mux-video';
+import type MuxVideoElement from '@mux/mux-video';
+import type { MediaError } from '@mux/mux-video';
 import type { AttributeTokenList } from './helpers';
 
 export type MuxPlayerProps = Partial<MuxVideoElement> & {
